Files ready to compile
Checked all files for errors and corrected. Determined LOADER file I was working with was for a different menu system (advanced). As it was not working I replaced with the version that matches the existing published game. Finally, found the correct copy of LOOKY.BAS! Game can now be compiled and tested.