First successful compile
Finally got all files to compile after finding community routines to match some of the QuickPackPro functions relied on by Baseball only. Graphics mode for the game has not even been tested, and there are no doubt bugs to work out for many modules. But this is a stage worth saving.